Правки по комментариям из пулл реквеста 4
