+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -5,7 +5,6 @@ import ru.yandex.practicum.filmorate.model.Film;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
 public interface FilmStorage {
     Collection<Film> findAll();
@@ -20,7 +19,9 @@ public interface FilmStorage {
 
     boolean deleteLike(Long filmId, Long userId);
 
-    Optional<Film> findFilmById(Long filmId);
+    Film findFilmById(Long filmId);
 
     List<Film> getTopFilms(int size);
+
+    boolean isExist(Long filmId);
 }
