+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryFilmStorage.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.storage;
 
 import lombok.Data;
 import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.Collection;
@@ -9,7 +10,6 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Component
@@ -27,8 +27,11 @@ public class InMemoryFilmStorage implements FilmStorage {
     }
 
     @Override
-    public Optional<Film> findFilmById(Long filmId) {
-        return Optional.ofNullable(films.get(filmId));
+    public Film findFilmById(Long filmId) {
+        if (!isExist(filmId)) {
+            throw new NotFoundException("Фильм по данному ID не найден");
+        }
+        return films.get(filmId);
     }
 
     @Override
@@ -58,5 +61,10 @@ public class InMemoryFilmStorage implements FilmStorage {
     @Override
     public boolean deleteLike(Long filmId, Long userId) {
         return films.get(filmId).getLikes().remove(userId);
+    }
+
+    @Override
+    public boolean isExist(Long filmId) {
+        return films.containsKey(filmId);
     }
 }
